# swagger.json
{
  "basePath": "/api/v1",
  "definitions": {
...
    "ConfirmationWithSender": {
      "allOf": [
        {
          "$ref": "#/definitions/Confirmation"
        },
        {
          "properties": {
            "from": {
              "format": "uuid",
              "type": "string"
            }
          },
          "required": ["from"],
          "type": "object"
        }
      ]
    },
    "ErrorMessage": {
      "properties": {
        "code": {
          "type": "integer"
        },
        "error": {
          "type": "string"
        }
      },
      "required": ["code", "error"],
      "type": "object"
    },
...
    "Instance": {
      "allOf": [
        {
          "$ref": "#/definitions/InstanceAndName"
        },
        {
          "properties": {
            "backend": {
              "type": "string"
            },
            "clientId": {
              "type": "string"
            }
          },
          "type": "object"
        }
      ]
    },
    "InstanceAndName": {
      "properties": {
        "instanceId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "Message": {
      "properties": {
        "confirmations": {
          "items": {
            "$ref": "#/definitions/ConfirmationWithSender"
          },
          "type": "array"
        },
        "content": {
          "$ref": "#/definitions/MessageContent"
        },
        "conversation": {
          "format": "uuid",
          "type": "string"
        },
        "from": {
          "format": "uuid",
          "type": "string"
        },
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "messageTimer": {
          "type": "integer"
        },
        "reactions": {
          "items": {
            "$ref": "#/definitions/ReactionWithSender"
          },
          "type": "array"
        },
        "state": {
          "enum": ["PayloadBundleState.INCOMING", "PayloadBundleState.OUTGOING_SENT"],
          "type": "string"
        },
        "timestamp": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "required": ["conversation", "from", "id", "messageTimer", "state", "timestamp", "type"],
      "type": "object"
    },
...
    "ServerErrorMessage": {
      "allOf": [
        {
          "$ref": "#/definitions/ErrorMessage"
        },
        {
          "properties": {
            "stack": {
              "type": "string"
            }
          },
          "type": "object"
        }
      ]
    },
...
  },
  "info": {
    "contact": {
      "email": "opensource@wire.com"
    },
    "description": "End-to-end Test Service (ETS) for Wire's test automation suite.",
    "license": {
      "name": "GPL-3.0",
      "url": "https://github.com/wireapp/wire-web-ets/blob/master/LICENSE"
    },
    "termsOfService": "https://wire.com/legal/",
    "title": "E2E Test Service",
    "version": "1.20.0"
  },
  "paths": {
...
    "/instance/{instanceId}": {
      "delete": {
        "description": "",
        "operationId": "deleteInstance",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Delete an instance",
        "tags": ["Instance"]
      },
      "get": {
        "operationId": "getInstance",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Instance"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Get information about an instance",
        "tags": ["Instance"]
      }
    },
    "/instance/{instanceId}/archive": {
      "post": {
        "operationId": "archiveConversation",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          },
          {
            "description": "",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "archive": {
                  "type": "boolean"
                },
                "conversationId": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/InstanceAndName"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Archive a conversation",
        "tags": ["Instance"]
      }
    },
...
    "/instance/{instanceId}/clear": {
      "post": {
        "operationId": "clearConversation",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          },
          {
            "description": "",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "conversationId": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/InstanceAndName"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Clear a conversation",
        "tags": ["Instance"]
      }
    },
...
    "/instance/{instanceId}/delete": {
      "post": {
        "operationId": "deleteMessage",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          },
          {
            "description": "",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "conversationId": {
                  "format": "uuid",
                  "type": "string"
                },
                "messageId": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/InstanceAndName"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Delete a message locally",
        "tags": ["Instance"]
      }
    },
    "/instance/{instanceId}/deleteEverywhere": {
      "post": {
        "operationId": "deleteMessageEverywhere",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          },
          {
            "description": "",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "conversationId": {
                  "format": "uuid",
                  "type": "string"
                },
                "messageId": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/InstanceAndName"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Delete a message for everyone",
        "tags": ["Instance"]
      }
    },
...
    "/instance/{instanceId}/getMessages": {
      "post": {
        "operationId": "getMessages",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          },
          {
            "description": "",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "conversationId": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "items": {
                "$ref": "#/definitions/Message"
              },
              "type": "array"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Get all messages",
        "tags": ["Instance"]
      }
    },
...
  },
  "schemes": ["http"],
  "swagger": "2.0",
...
}